feat: Use modules to segregate the examples
I've decided to use modules to keep the example funtion names short and
have the module namespaces track the organization of the
Plotters Developers Guide.

So, here I've made module `getting_started` with `sin_wave()` and module
`basic_plotting` with the first sub-module `define_chart_context` which
has two functions `chart_builder()` and `draw_series()`. Which matches
the organization of the guide.